Remove old CLI workspace sidebar
