+++ apps/sqlrooms-cli-ui/src/workspace/sidebar/CliAssistantToggleButton.tsx
@@ -1,0 +1,29 @@
+import {Button, Tooltip, TooltipContent, TooltipTrigger} from '@sqlrooms/ui';
+import {SparklesIcon} from 'lucide-react';
+import {useRoomStore} from '../../store';
+
+export function CliAssistantToggleButton() {
+  const toggleCollapsed = useRoomStore((state) => state.layout.toggleCollapsed);
+  const isAssistantCollapsed = useRoomStore((state) =>
+    state.layout.isCollapsed('assistant-sidebar'),
+  );
+
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>
+        <Button
+          type="button"
+          variant="ghost"
+          size="icon"
+          className="text-muted-foreground hover:text-foreground data-[active=true]:bg-accent data-[active=true]:text-foreground size-9"
+          onClick={() => toggleCollapsed('assistant-sidebar')}
+          data-active={!isAssistantCollapsed}
+        >
+          <SparklesIcon className="h-4 w-4" />
+          <span className="sr-only">AI Assistant</span>
+        </Button>
+      </TooltipTrigger>
+      <TooltipContent>AI Assistant</TooltipContent>
+    </Tooltip>
+  );
+}
