+++ apps/sqlrooms-cli-ui/src/workspace/sidebar/CliSidebarToggleButton.tsx
@@ -1,0 +1,23 @@
+import {
+  SidebarTrigger,
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+  useSidebar,
+} from '@sqlrooms/ui';
+
+export function CliSidebarToggleButton() {
+  const {state} = useSidebar();
+
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>
+        <SidebarTrigger
+          className="text-muted-foreground hover:text-foreground data-[active=true]:bg-accent data-[active=true]:text-foreground size-9"
+          data-active={state === 'expanded'}
+        />
+      </TooltipTrigger>
+      <TooltipContent>Toggle sidebar</TooltipContent>
+    </Tooltip>
+  );
+}
