+++ apps/sqlrooms-cli-ui/src/workspace/sidebar/index.ts
@@ -1,0 +1,3 @@
+export {CliAssistantToggleButton} from './CliAssistantToggleButton';
+export {CliSidebarToggleButton} from './CliSidebarToggleButton';
+export {CliWorkspaceSidebar} from './CliWorkspaceSidebar';
