+++ apps/sqlrooms-cli-ui/src/workspace/sidebar/useCliArtifactSidebarTabs.ts
@@ -1,0 +1,71 @@
+import {createArtifactLayoutNode} from '@sqlrooms/artifacts';
+import {useCallback, useMemo} from 'react';
+import {CLI_ARTIFACT_TYPES, type CliArtifactType} from '../../artifactTypeIds';
+import {useRoomStore} from '../../store';
+
+export function useCliArtifactSidebarTabs() {
+  const artifactsConfig = useRoomStore((state) => state.artifacts.config);
+  const artifactTypes = useRoomStore((state) => state.artifacts.artifactTypes);
+  const aiSessions = useRoomStore((state) => state.ai.config.sessions);
+  const aiSessionArtifacts = useRoomStore(
+    (state) => state.artifactAi.config.aiSessionArtifacts,
+  );
+  const selectedTabId = useRoomStore((state) =>
+    state.layout.getActiveTab('workspace'),
+  );
+  const addTab = useRoomStore((state) => state.layout.addTab);
+  const setActiveTab = useRoomStore((state) => state.layout.setActiveTab);
+  const setCurrentArtifact = useRoomStore(
+    (state) => state.artifacts.setCurrentArtifact,
+  );
+  const setShowArtifactChooser = useRoomStore(
+    (state) => state.workspaceUi.setShowArtifactChooser,
+  );
+
+  const runningSessionCountsByArtifact = useMemo(() => {
+    const counts: Record<string, number> = {};
+    for (const session of aiSessions) {
+      if (!session.isRunning) continue;
+      const artifactId = aiSessionArtifacts[session.id];
+      if (!artifactId) continue;
+      counts[artifactId] = (counts[artifactId] ?? 0) + 1;
+    }
+    return counts;
+  }, [aiSessionArtifacts, aiSessions]);
+
+  const tabs = useMemo(
+    () =>
+      artifactsConfig.artifactOrder
+        .map((artifactId) => artifactsConfig.artifactsById[artifactId])
+        .filter((artifact) => {
+          return (
+            artifact &&
+            CLI_ARTIFACT_TYPES.includes(artifact.type as CliArtifactType) &&
+            artifact.visibility === 'workspace'
+          );
+        })
+        .map((artifact) => ({
+          id: artifact.id,
+          name: artifact.title,
+          type: artifact.type,
+          runningSessionCount: runningSessionCountsByArtifact[artifact.id] ?? 0,
+        })),
+    [
+      artifactsConfig.artifactOrder,
+      artifactsConfig.artifactsById,
+      runningSessionCountsByArtifact,
+    ],
+  );
+
+  const selectArtifact = useCallback(
+    (artifactId: string) => {
+      addTab('workspace', createArtifactLayoutNode(artifactId, 'artifact'));
+      setActiveTab('workspace', artifactId);
+      setCurrentArtifact(artifactId);
+      setShowArtifactChooser(false);
+    },
+    [addTab, setActiveTab, setCurrentArtifact, setShowArtifactChooser],
+  );
+
+  return {artifactTypes, selectedTabId, selectArtifact, tabs};
+}
